ObjectPool Deleter Extraction
It became desirable to extract the nested ObjectPool Deleter class in order to make declaring createObj function pointers more straight forward for users of ObjectPool. As it were, this was all but impossible due to a default template parameter of ObjectPool. It was not desirable to remove  the default parameter. The new ObjectPoolDeleter has no dependence on concrete ObjectPool template instantiations. It depends on the generic ObjectPoolBase instead.