ObjectPool Template Signature Change
This change to remove the template parameter for minTypeSizeAlloc which defaulted to the sizeof base type, was necessary in order to forward declare object pools. We loose compile time size overrun errors for runtime size overrun errors but we gain utility. ObjectPool is something that will be interfaced by multiple components.