MessageQueue Dispatch Lock Work Complete
I wanted to test the getAutoDispatchLock feature and it is a good thing that I did because, it was bugged. I had made the assumption that default move constructor and assignment operator would do the correct swapping of null details. However, that was not happenning because details is a raw pointer. I had to provide implementations for them to do the appropriate thing.

This work is complete.