Dispatch Locking Is Now Optional
Supporting dispatch locking imposes an extra mutex
lock/release in an implementation's dispatch loop. It should be the client's option as to whether this is needed for its synchronization mechanisms on a case by case basis. This has now been implemented. Dispatch Locking support is disabled by default and may be enabled during construction of MessageQueue.