Reorganize Semaphore Implementation
Reorganize the Semaphore implementation to avoid CTidy complaints plus some attribute renaming for the next possible attempt to make the Semaphore bi-polar (block on notify limit). This was attempted before and abandoned. I am going to attempt to resurrect it but may not be successful. I want to at least capture these changes.