Semaphore Bipolar Capability Tested
It sure seems like I finally landed this one. As if now, this Semaphore not only blocks when no available counts are left to take. It also blocks when a maximum available count would be exceeded on a give. Wow, who ever heard of a bipolar semaphore before. Perhaps one of us needs therapy.