MessageQueue Documentation Plus
MessageQueue lacked a custom exception if getAutoDispatchLock is invoked and dispatch locking was not enabled during construction. The documentation was also not current with exception types thrown. This is a very subtle API change on getAutoDispatchLock regarding what it might throw, but we are merely throwing a subclass of what was originally thrown. Insignificant JobDispatcher Tweaks.