Provide Mechanism to Determine BlockPool Array Size
A weakness of std::unique_ptr< T[] > is that you cannot determine the size of the array. BlockPool creates std::unique_ptr< T[] > with a custom deleter. We've modified this deleter to provide a method to query the number of elements. We have also cleaned up MemoryPool base code for proper const-ness.